WEBSITE-UGT/BACK-END/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
python-dotenv>=1.0.0
gspread>=6.1.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
slowapi>=0.1.9
email-validator>=2.2.0 